v1/galion/services/auth-service/requirements.txt
# FastAPI framework - Updated to latest secure version
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database - Latest secure versions
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication - Latest secure versions
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Redis
redis==5.0.1

# Kafka
kafka-python==2.0.2

# Security enhancements
python-dotenv==1.0.0
slowapi==0.1.9  # Rate limiting
python-json-logger==2.0.7  # Structured logging

# 2FA/TOTP support (GitHub-style 2FA)
pyotp==2.9.0  # Time-based OTP
qrcode[pil]==7.4.2  # QR code generation
Pillow==10.2.0  # Image processing for QR codes

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Monitoring and security
prometheus-client==0.19.0
